backend/requirements.txt
# Requirements for backend environment
# (cleaned: use # for comments and one package per line)
base64io
deep_translator
deepgram-sdk
fastapi
groq
hf_xet
langchain
langchain-chroma
langchain-core
langchain-groq
langchain-huggingface
langchain-nvidia-ai-endpoints
langchain_community
langgraph
numpy
pdf2image
pdfminer.six
pi-heif
PyPDF2
python-dotenv
python-magic
pytesseract
resemblyzer
requests
tiktoken
unstructured
unstructured_inference
unstructured_pytesseract
uvicorn